+++ oeving2/track.h
@@ -1,6 +1,8 @@
 #ifndef TRACK_H
 #define TRACK_H
 
+#include <inttypes.h>
+
 #include "sound.h"
 
 #define TRACKS_TOTAL 8
@@ -8,9 +10,9 @@
 // Structure for a sound and playback information
 typedef struct {
     sound_t *sound;
-    int current_sample;
-    int current_sample_iteration;
-    int current_sample_point;
+    int16_t current_sample;
+    int16_t current_sample_iteration;
+    int16_t current_sample_point;
 } track_t;
 
 track_t **tracks;
@@ -19,6 +21,6 @@ track_t **tracks;
 void tracks_init();
 track_t* track_new();
 void track_play(track_t *track, sound_t *sound);
-int track_advance(track_t* track);
+int32_t track_advance(track_t* track);
 
 #endif
